pyproject.toml
[project]
name = "churn-detection"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "mlflow>=3.2.0"
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]   # errors, formatting, bugbear, imports
ignore = ["E501"]               # example: ignore line length

[tool.ruff.format]
quote-style = "double"
indent-style = "space"